Refactor: create separate downloader module, move OS again
