+++ engine/launcher/src/main/scala/org/enso/launcher/upgrade/LauncherUpgrader.scala
@@ -3,8 +3,8 @@ package org.enso.launcher.upgrade
 import java.nio.file.{Files, Path}
 import com.typesafe.scalalogging.Logger
 import nl.gn0s1s.bump.SemVer
-import org.enso.cli.CLIOutput
-import org.enso.distribution.{DistributionManager, FileSystem, OS}
+import org.enso.cli.{CLIOutput, OS}
+import org.enso.distribution.{DistributionManager, FileSystem}
 import org.enso.distribution.locking.{
   LockType,
   LockUserInterface,
@@ -13,7 +13,7 @@ import org.enso.distribution.locking.{
 }
 import org.enso.runtimeversionmanager.CurrentVersion
 import org.enso.distribution.FileSystem.PathSyntax
-import org.enso.runtimeversionmanager.archive.Archive
+import org.enso.downloader.archive.Archive
 import org.enso.runtimeversionmanager.components.UpgradeRequiredError
 import org.enso.launcher.cli.{
   CLIProgressReporter,
