+++ engine/launcher/src/test/scala/org/enso/launcher/installation/InstallerSpec.scala
@@ -1,9 +1,10 @@
 package org.enso.launcher.installation
 
-import org.enso.distribution.{FileSystem, OS}
+import org.enso.distribution.FileSystem
 
 import java.nio.file.{Files, Path}
 import FileSystem.PathSyntax
+import org.enso.cli.OS
 import org.enso.runtimeversionmanager.test.WithTemporaryDirectory
 import org.enso.launcher._
