+++ engine/launcher/src/test/scala/org/enso/launcher/upgrade/UpgradeSpec.scala
@@ -3,9 +3,10 @@ package org.enso.launcher.upgrade
 import java.nio.file.{Files, Path, StandardCopyOption}
 import io.circe.parser
 import nl.gn0s1s.bump.SemVer
-import org.enso.distribution.{FileSystem, OS}
+import org.enso.distribution.FileSystem
 import org.enso.distribution.locking.{FileLockManager, LockType}
 import FileSystem.PathSyntax
+import org.enso.cli.OS
 import org.enso.launcher._
 import org.enso.runtimeversionmanager.test.WithTemporaryDirectory
 import org.enso.testkit.RetrySpec
