+++ lib/scala/distribution-manager/src/main/scala/org/enso/distribution/Environment.scala
@@ -1,6 +1,7 @@
 package org.enso.distribution
 
 import com.typesafe.scalalogging.Logger
+import org.enso.cli.OS
 import org.enso.logger.masking.MaskedString
 
 import java.io.File
