+++ lib/scala/downloader/src/main/scala/org/enso/downloader/archive/ArchiveFormat.scala
@@ -1,4 +1,4 @@
-package org.enso.runtimeversionmanager.archive
+package org.enso.downloader.archive
 
 import java.nio.file.Path
