+++ lib/scala/runtime-version-manager-test/src/test/scala/org/enso/runtimeversionmanager/components/RuntimeVersionManagerSpec.scala
@@ -2,7 +2,8 @@ package org.enso.runtimeversionmanager.components
 
 import java.nio.file.{Files, Path}
 import nl.gn0s1s.bump.SemVer
-import org.enso.distribution.{FileSystem, OS}
+import org.enso.cli.OS
+import org.enso.distribution.FileSystem
 import org.enso.distribution.FileSystem.PathSyntax
 import org.enso.runtimeversionmanager.config.GlobalConfigurationManager
 import org.enso.runtimeversionmanager.releases.ReleaseNotFound
